Fixed floating poing division error

+++ lib/frequencyprocessing.c
@@ -1,7 +1,7 @@
 #include "frequencyprocessing.h"
 
 struct signal keepPositiveFreq(struct signal data){
-    data.delta_f = data.fs/data.length;
+    data.delta_f = (float) data.fs / (float) data.length;
     data.length = data.length/2;
     uint32_t j = 0;
     for(uint32_t i=0; i<data.length; i++){
